+++ packages/tolk/lib/uop/uop.ml
@@ -73,6 +73,7 @@ type queue_info = {
   table : int;
   inputs : (int * string) list;
   outputs : int list;
+  timings : (string * int * int * int) list;
   independent_accesses : (int * int) list;
   accesses : int list list;
 }
@@ -3662,7 +3663,7 @@ let to_elf u =
   | _ -> invalid_arg "Uop.to_elf: expected a compiled PROGRAM"
 
 let export_magic = "TOLKUOP\x00"
-let export_version = 30
+let export_version = 31
 
 type serialized_node = {
   serialized_op : Ops.t;